Shelf W and H are UNSIGNED INT, not double ._.

+++ src/APC/Shelf.cpp
@@ -10,8 +10,8 @@ namespace APC {
   const Pose Shelf::POSE(0.792, 0.454, 0.836, 0, 0, 0);
 
   /** Size of the shelf in #bins */
-  const double Shelf::HEIGHT=4;
-  const double Shelf::WIDTH=3;
+  const unsigned int Shelf::HEIGHT=4;
+  const unsigned int Shelf::WIDTH=3;
 
   /** Size of a single bin */
   const double Shelf::BIN_HEIGHT=0.3;
